chore : redone exercises
J'ai refait tout les premiers exercices

+++ exercises/02_functions/functions4.rs
@@ -3,12 +3,12 @@
 // Don't worry about the function bodies themselves, we are only interested in
 // the signatures for now.
 
-fn is_even(num: i64) -> bool {
+fn is_even(num: i8) -> bool {
     num % 2 == 0
 }
 
 // TODO: Fix the function signature.
-fn sale_price(price: i64) -> i64 {
+fn sale_price(price: i8) -> i8 {
     if is_even(price) {
         price - 10
     } else {
